Replace private_config with .jfsenv
In practice, this is functionally identical. We're sourcing some config
outside of source control so we can do private stuff.

But there are a few improvements here:

- We keep `.jfsenv` at the root of a system path that's guaranteed to
exist - as opposed to keeping it stashed in the `dotfiles` repo, which
then requires some kind of linking process post-clone. Ultimately we
still need to maintain `.jfsenv` or a symlink to `~/Documents/.jfsenv`,
but this makes more sense than injecting the privately maintained
dotfile into the `dotfiles` repo.
- This is semantically more appropriate. We were overloading
`private_config` with envars and dynamic config, which turned it into a
bit of a beast. `.jfsenv` is strictly used for establishing environment
context with any hardcoded private paths. This then serves as a central
context manifest, which other scripts can then consume - instead of
having to orchestrate all this in `private_config`.